Update index.js
Join Txon as a Web Development and Design Intern, where you'll have the opportunity to work on exciting projects, enhance your skills in modern web technologies, and collaborate with a dynamic team. As an intern, you'll be involved in designing and developing responsive websites, implementing user-friendly interfaces, and ensuring seamless functionality across various devices. This role is perfect for passionate individuals eager to gain hands-on experience in a fast-paced and innovative environment. Apply now to kickstart your career in web development and design with Txon!